ThreadDumpCompare throws exception on first access #14
seems that there is a race condition in AnalyzerRoute#findOrAllocate if
multiple requests with the same parameters are made within a short
timeframe.
In that case, the 'insert' part of allocate fails because of a
key violation. findOrAllocate now does a retry. If the insert fails, it
should find a running job in the next try.

closes #14